Move AnimationRules into declaration_block.rs.
